+++ src/index.d.ts
@@ -68,8 +68,6 @@ export function disposeOnUnmount<TF extends Disposer | Disposer[]>(
 /**
  * Utilities
  */
-export function onError(cb: (error: Error) => void): () => void
-
 export class Provider extends React.Component<any, {}> {}
 
 export class Observer extends React.Component<
@@ -81,35 +79,6 @@ export class Observer extends React.Component<
 > {}
 
 export function useStaticRendering(value: boolean): void
-
-/**
- * Enable dev tool support, makes sure that renderReport emits events.
- */
-export function trackComponents(): void
-
-export const renderReporter: RenderReporter
-
-export interface RenderReporter {
-    on(handler: (data: IRenderEvent) => void): void
-}
-
-export interface IRenderEvent {
-    event: "render" | "destroy"
-    renderTime?: number
-    totalTime?: number
-    component: React.ReactElement<any> // Component instance
-    node: any // DOMNode
-}
-
-/**
- * WeakMap DOMNode -> Component instance
- * @deprecated
- */
-export const componentByNodeRegistery: any
-/**
- * WeakMap DOMNode -> Component instance
- */
-export const componentByNodeRegistry: any
 
 /**
  * @deprecated, use PropTypes instead