+++ test/index.js
@@ -5,6 +5,15 @@ import ReactDOM from "react-dom"
 
 configure({ adapter: new Adapter() })
 
+// Uglyness to find missing 'act' more easily
+// 14-2-19 / React 16.8.1, temporarily work around, as error message misses a stack-trace
+Error.stackTraceLimit = Infinity
+const origError = console.error
+console.error = function(msg) {
+    if (/react-wrap-tests-with-act/.test("" + msg)) throw new Error("missing act")
+    return origError.apply(this, arguments)
+}
+
 export function createTestRoot() {
     if (!window.document.body) {
         window.document.body = document.createElement("body")